requirements.txt
django==3.0
django-ckeditor==5.2.1
django-crispy-forms==1.6.1
django-recaptcha==1.2.1
django-cleanup
Pillow